Moved files logic into dedicated module.
