+++ scripts/js/user_modules/files_helper.js
@@ -1,0 +1,16 @@
+const { resolve } = require('path');
+const { readdir } = require('fs').promises;
+
+async function* getFiles(dir) {
+    const dirents = await readdir(dir, { withFileTypes: true });
+    for (const dirent of dirents) {
+      const res = resolve(dir, dirent.name);
+      if (dirent.isDirectory()) {
+        yield* getFiles(res);
+      } else {
+        yield res;
+      }
+    }
+  }
+
+module.exports = { getFiles }
